+++ pkg/bitesize/service.go
@@ -15,6 +15,7 @@ import (
 type Service struct {
 	Name         string                  `yaml:"name" validate:"nonzero"`
 	ExternalURL  []string                `yaml:"-"`
+	Backend      string                  `yaml:"backend"`
 	Ports        []int                   `yaml:"-"` // Ports have custom unmarshaler
 	Ssl          string                  `yaml:"ssl,omitempty" validate:"regexp=^(true|false)*$"`
 	Version      string                  `yaml:"version,omitempty"`