+++ pkg/cluster/service_map.go
@@ -160,6 +160,8 @@ func (s ServiceMap) AddIngress(ingress v1beta1_ext.Ingress) {
 	biteservice.HTTPSBackend = httpsBackend
 	biteservice.HTTPSOnly = httpsOnly
 	biteservice.Ssl = ssl
+
+	biteservice.Backend = ingress.Spec.Rules[0].IngressRuleValue.HTTP.Paths[0].Backend.ServiceName
 }
 
 func (s ServiceMap) AddMongoStatefulSet(statefulset v1beta1_apps.StatefulSet) {